refactor: move extension methods and tests related to rendering process to their own files
